Pull spawn_file_watcher into its own function
